Refactor: Move tests to separate files and create common test module
